requirements.txt
# local package
-e .
requests~=2.28.1
pandas~=2.0.1
nbformat~=5.8.0
nbconvert~=7.4.0
numpy~=1.24.3
pingouin
python-dotenv
ipython
radon
GitPython
tqdm
seaborn
scipy
setuptools~=65.5.0